refactor[api-server]: 통합테스트 전용 패키지 생성
