+++ src/test/java/com/watermelon/server/integration/PartsIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.event.lottery;
+package com.watermelon.server.integration;
 
 import com.watermelon.server.BaseIntegrationTest;
 import com.watermelon.server.event.lottery.auth.service.TestTokenVerifier;
@@ -10,7 +10,6 @@ import com.watermelon.server.event.lottery.parts.repository.LotteryApplierPartsR
 import com.watermelon.server.event.lottery.parts.repository.PartsRepository;
 import com.watermelon.server.event.lottery.repository.LotteryApplierRepository;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,9 +17,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import static com.watermelon.server.Constants.*;
 import static com.watermelon.server.common.constants.PathConstants.PARTS;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@DisplayName("[통합] 추첨 통합 테스트")
-public class LotteryIntegrationTest extends BaseIntegrationTest {
+@DisplayName("[통합] 파츠 통합 테스트")
+public class PartsIntegrationTest extends BaseIntegrationTest {
 
     @Autowired
     private LotteryApplierRepository lotteryApplierRepository;
@@ -28,21 +28,51 @@ public class LotteryIntegrationTest extends BaseIntegrationTest {
     @Autowired
     private PartsRepository partsRepository;
 
-    @BeforeEach
-    void setUp() {
-        partsRepository.saveAll(Parts.createAllParts());
-    }
+    @Autowired
+    private LotteryApplierPartsRepository lotteryApplierPartsRepository;
 
     @Test
-    @DisplayName("추첨 이벤트 응모 - 성공")
-    void applySuccessTest() throws Exception {
+    @DisplayName("추첨권 부여 - 성공")
+    void partsLotteryPolicyTest() throws Exception {
+
+        //given
+        LotteryApplier lotteryApplier = saveTestLotteryApplier();
+        for(PartsCategory partsCategory:PartsCategory.values()){
+            Parts parts = partsRepository.save(Parts.createTestCategoryParts(partsCategory));
+            lotteryApplierPartsRepository.save(
+                    LotteryApplierParts.createApplierParts(true, lotteryApplier, parts)
+            );
+        }
 
         //when
         postPartsLotteryRequestTestUser();
+        boolean expected = findTestLotteryApplier().isPartsApplier();
 
         //then
-        LotteryApplier lotteryApplier = findTestLotteryApplier();
-        Assertions.assertThat(lotteryApplier.isLotteryApplier()).isTrue();
+        Assertions.assertThat(expected).isTrue();
+
+    }
+
+    @Test
+    @DisplayName("파츠 뽑기 - 뽑기권 없음")
+    void partsLotteryNotRemainCountExceptionTest() throws Exception {
+
+        //given
+        postPartsLotteryRequestTestUser();
+
+        //when
+        mvc.perform(post(PARTS)
+                .header( HEADER_NAME_AUTHORIZATION, HEADER_VALUE_BEARER+HEADER_VALUE_SPACE+TestTokenVerifier.TEST_VALID_TOKEN)
+        )
+                //then
+                .andExpect(status().isTooManyRequests());
+
+    }
+
+    private LotteryApplier saveTestLotteryApplier() {
+        LotteryApplier lotteryApplier = LotteryApplier.createLotteryApplier(TestTokenVerifier.TEST_UID);
+        lotteryApplierRepository.save(lotteryApplier);
+        return lotteryApplier;
     }
 
     private LotteryApplier findTestLotteryApplier() {
